+++ packages/cdai/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.3](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive-cdai@1.3.2...@carbon/ibm-cloud-cognitive-cdai@1.3.3) (2021-06-17)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive-cdai
+
+
+
+
+
 ## [1.3.2](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive-cdai@1.3.1...@carbon/ibm-cloud-cognitive-cdai@1.3.2) (2021-06-10)
 
 **Note:** Version bump only for package @carbon/ibm-cloud-cognitive-cdai