+++ packages/cloud-cognitive/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.45.3](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.45.2...@carbon/ibm-cloud-cognitive@0.45.3) (2021-06-17)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+## [0.45.2](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.45.1...@carbon/ibm-cloud-cognitive@0.45.2) (2021-06-16)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+## [0.45.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.45.0...@carbon/ibm-cloud-cognitive@0.45.1) (2021-06-15)
+
+
+### Bug Fixes
+
+* **HTTPErrors:** remove canary check on http error content component ([#900](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/900)) ([e6e394a](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/e6e394a4e90d3bfda58489a9a5687e0499215fb9))
+
+
+
+
+
+# [0.45.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.44.0...@carbon/ibm-cloud-cognitive@0.45.0) (2021-06-15)
+
+
+### Features
+
+* breadcrumb overflow title and href updates ([#898](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/898)) ([290198b](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/290198bdd078060f4a91bb6c27904aa17ce21b7b))
+
+
+
+
+
+# [0.44.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.43.1...@carbon/ibm-cloud-cognitive@0.44.0) (2021-06-15)
+
+
+### Features
+
+* simplified carbon token value in js ([#894](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/894)) ([a225eee](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/a225eee68364554f62f1f41889164c3f0e48e645))
+
+
+
+
+
+## [0.43.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.43.0...@carbon/ibm-cloud-cognitive@0.43.1) (2021-06-15)
+
+
+### Bug Fixes
+
+* page header performance glitches ([#896](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/896)) ([57eae46](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/57eae4653f3b9efbe1218ad947ff6141760f0c14))
+
+
+
+
+
 # [0.43.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.42.9...@carbon/ibm-cloud-cognitive@0.43.0) (2021-06-11)